fix(input_format): match datatypes with official `mlsys.h` definitions
I doubt that this matters but just to be safe

+++ src/input_format.rs
@@ -8,13 +8,13 @@ pub enum OperationType {
 
 #[derive(Clone, Debug, Deserialize)]
 pub struct InputFormat {
-    pub widths: Vec<usize>,
-    pub heights: Vec<usize>,
+    pub widths: Vec<i64>,
+    pub heights: Vec<i64>,
     pub inputs: Vec<Vec<usize>>,
     pub outputs: Vec<Vec<usize>>,
-    pub base_costs: Vec<u32>,
+    pub base_costs: Vec<i64>,
     pub op_types: Vec<OperationType>,
-    pub fast_memory_capacity: usize,
-    pub slow_memory_bandwidth: i32,
-    pub native_granularity: (usize, usize),
+    pub fast_memory_capacity: i64,
+    pub slow_memory_bandwidth: i64,
+    pub native_granularity: (i64, i64),
 }
